backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django_csp==3.8
djangorestframework==3.15.2
drf-yasg==1.21.7
eval_type_backport==0.2.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
inflection==0.5.1
jsonpath-python==1.0.6
mistralai==1.1.0
mypy-extensions==1.0.0
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.8.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
